% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latentSNA.R
\name{latentSNA}
\alias{latentSNA}
\title{Attribute informed brain connectivity}
\usage{
latnetSNA(X, Y, W, H, seed = 1, nscan =
10000, burn = 500, odens = 25,
prior=list())
}
\arguments{
\item{X}{a list of V x V brain connectivity data.}

\item{Y}{a matrix of N x P individual outcome data.}

\item{W}{a matrix of N x Q covariates for the connectivity data.}

\item{H}{a matrix of N x Q1 covariates for the attribute data.}

\item{seed}{random seed}

\item{nscan}{number of iterations of the Markov chain (beyond burn-in)}

\item{burn}{burn in for the Markov chain}

\item{odens}{output density for the Markov chain, chain thinning every odens iterations}

\item{prior}{list: A list of hyperparameters for the prior distribution}
}
\value{
\item{COV}{posterior mean of the covariance parameters between brain and behaviors}
\item{BETAPM}{posterior mean of the regression coefficient parameters for the connectivity data}
\item{GAMMAPM}{posterior mean of the regression coefficient parameters for the attribute data}
\item{THETAPM}{posterior mean of the latent person variable}
\item{APM}{posterior mean of connectivity intercepts} \item{BPM}{posterior
mean of attribute intercepts} \item{U}{last iteration of latent connectivity for all regions
} 
\item{UPM}{posterior mean of U}
\item{UVPM.l}{list of posterior mean of connectivity}
\item{Theta}{the last iteration of the Theta estimate}
\item{X}{observed X}
\item{Y}{observed Y}
\item{EFlPM}{posterior mean estimates of X}
 \item{ETPM}{posterior mean estiamtes of Y}
\item{TMPM}{posterior mean estimates of latent behavior component}
 \item{FLPM}{posterior mean estiamtes of latent connectivity component}
\item{input}{input values}
}
\description{
An MCMC algorithm for fitting the latentSNA model
}
\examples{
attach(X)
attach(Y)

## More MCMC burn-in iterations and iterations are needed than specified here. 
model1=latentSNA(X, Y,W=NULL, H=NULL,
indices = NULL, indices_irt = NULL,
seed = 1, nscan = 1, burn = 1, odens = 1,
print = TRUE, gof=TRUE, plot=TRUE,
prior=list())
}
\author{
Selena Wang
}
