% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latentSNA_PACKAGE.R
\docType{package}
\name{latentSNA_PACKAGE}
\alias{LatentSNA}
\alias{LatentSNA-package}
\alias{latentSNA_PACKAGE}
\alias{latentSNA}
\title{Create LatentSNA model for brain-behavior linking}
\description{
\code{latentSNA} provides a latent model for jointly modeling
brain connectivity with individual outcomes. The latent space models are implemented using the
MCMC inference approach.
}
\details{
LatentSNA: the function \code{\link{latentSNA}} implements LatentSNA introduced by
Wang et.al (2023).
}
\references{
arXiv preprint arXiv:2309.11349, 2023.
}
