% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbeta_a.R
\name{rbeta_a}
\alias{rbeta_a}
\title{Conditional simulation of intercept and regression coefficients}
\usage{
rbeta_a(Fl, W = NULL, s2 = 1, U = U, ivA = NULL, beta0 = NULL, S0 = NULL)
}
\arguments{
\item{Fl}{a list of V X V normal connectivity matrix}

\item{W}{N x Q covariate matrix}

\item{s2}{variance}

\item{U}{a V by V by N array}

\item{ivA}{prior inverse variance for the intercept parameters}

\item{beta0}{prior mean vector for regression parameters}

\item{S0}{prior precision matrix for regression parameters}
}
\value{
\item{beta}{regression coefficients} \item{a}{subject-specific intercept}
}
\description{
Simulates from the joint full conditional distribution of (beta)
in a brain connectivity model
}
\author{
Selena Wang
}
