% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rFl_nrm.R
\name{rFl_nrm}
\alias{rFl_nrm}
\title{Simulate missing values in a normal connectivity model}
\usage{
rFl_nrm(Z, EZ, s2, X)
}
\arguments{
\item{Z}{a square matrix, the current value of Z}

\item{EZ}{expected value of Z}

\item{s2}{dyadic variance}

\item{X}{square relational matrix}
}
\value{
a square matrix, equal to  at non-missing values
}
\description{
Simulates missing values of a sociomatrix under a normal connectivity model
}
\author{
Selena Wang
}
