% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rH_bin.R
\name{rH_bin}
\alias{rH_bin}
\title{Simulate H}
\usage{
rH_bin(H,EH,Y,s1=1)
}
\arguments{
\item{H}{a n X m matrix, the current value of H}

\item{EH}{expected value of H}

\item{Y}{n X m binary item response matrix}

\item{s1}{item response variance}
}
\value{
a n X m matrix, the new value of H
}
\description{
Simulates a random latent matrix H given its expectation
and a behavior matrix Y
}
\author{
Selena Wang
}
