% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rH_nrm.R
\name{rH_nrm}
\alias{rH_nrm}
\title{Simulate missing values in a normal behavior model}
\usage{
rH_nrm(H, EH,s1, Y)
}
\arguments{
\item{H}{a matrix, the current value of H}

\item{EH}{expected value of H}

\item{s1}{behavior variance}

\item{Y}{behavior matrix}
}
\value{
a behavior matrix, equal to  at non-missing values
}
\description{
Simulates missing values under a behavior model
}
\author{
Selena Wang
}
