% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs2.R
\name{rs2}
\alias{rs2}
\title{Gibbs update for connectivity variance}
\usage{
rs2(Fl, offset=0,nu2=NULL,s20=NULL)
}
\arguments{
\item{Fl}{a list of V X V normal connectivity matrix}

\item{nu2}{prior degrees of freedom}

\item{s20}{prior estimate of s2}
}
\value{
a new value of s2
}
\description{
Gibbs update for connectivity variance
}
\author{
Selena Wang
}
