% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rSu.R
\name{rSu}
\alias{rSu}
\title{Gibbs update for latent effects covariance}
\usage{
rSu(U,Su0=NULL,etau=NULL)
}
\arguments{
\item{U}{latent connectivity and behavior}

\item{Su0}{prior (inverse) scale matrix for the prior distribution}

\item{etau}{prior degrees of freedom for the prior distribution}
}
\description{
Gibbs update for latent effects covariance
}
\author{
Selena Wang
}
