% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rTheta_b.R
\name{rTheta_b}
\alias{rTheta_b}
\title{Gibbs sampling of Theta}
\usage{
rTheta_b(H, beta, Alpha, Theta ,U, Stheta, Sutheta, Su, s1)
}
\arguments{
\item{H}{N X P normal behavior matrix}

\item{beta}{P X 1 behavior intercept vector}

\item{Alpha}{1 X 1 intercept vector}

\item{Theta}{current value of Theta}

\item{U}{matrix containing current value of U}

\item{Stheta}{covariance of Theta}

\item{Sutheta}{covariance between U and Theta}

\item{Su}{matrix containing covariance of U}

\item{s1}{behavior variance}
}
\value{
\item{Theta}{a new value of Theta}
}
\description{
A Gibbs sampler for updating the Person latent effect Theta.
}
\author{
Selena Wang
}
