% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rU.R
\name{rU}
\alias{rU}
\title{Gibbs sampling of U}
\usage{
rU(Fl,U,Theta, Stheta, Sutheta, Su, s2=1, offset=offset)
}
\arguments{
\item{Fl}{a list of V X V normal relational matrix}

\item{EFl}{a list of the same dimension as Fl. It is assumed that
Fl-offset follows a SRRM, so the offset should contain any multiplicative
effects (such as \code{U\%*\% t(U) } )}

\item{U}{V X K matrix containing current value of U}

\item{Theta}{D X V current value of Theta}

\item{Stheta}{D X D covariance of Theta}

\item{Sutheta}{D X K covariance between U and Theta}

\item{Su}{K X K matrix containing covariance of U}

\item{s2}{dyadic variance}
}
\value{
\item{U}{a new value of U}
}
\description{
A Gibbs sampler for updating U.
}
\author{
Selena Wang
}
