% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rXi.R
\name{rXi}
\alias{rXi}
\title{Gibbs sampling of behaivor parameters Xi}
\usage{
rXi(H, beta, Alpha, Theta, mxi, Sigmaxi, s1 = 1)
}
\arguments{
\item{H}{normal behavior matrix}

\item{beta}{behavior intercept vector}

\item{Alpha}{behavior intercept vector}

\item{Theta}{current value of Theta}

\item{mxi}{vector of prior for the mean of Xi}

\item{Sigmaxi}{matrix of prior for the variance of Xi}

\item{s1}{behavior variance}
}
\value{
\item{beta}{a new value of beta} \item{Alpha}{a new value of Alpha}
}
\description{
A Gibbs sampler for updating the behavior parameters.
}
\author{
Selena Wang
}
