% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simH.R
\name{simH}
\alias{simH}
\title{Simulate H given its expectation and covariance}
\usage{
simH(EH, s1 = 1)
}
\arguments{
\item{EH}{expected value of H}

\item{s1}{attribute variance}
}
\value{
a simulated value of H
}
\description{
Simulate H given its expectation and covariance
}
\author{
Selena Wang
}
