% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simX_nrm.R
\name{simX_nrm}
\alias{simX_nrm}
\title{Simulate a normal connectivity matrix}
\usage{
simX_nrm(EX, s2)
}
\arguments{
\item{EX}{square matrix giving the expected value of the connectivity matrix}

\item{s2}{dyadic variance}
}
\value{
a square matrix
}
\description{
Simulates a normal connectivity matrix
}
\author{
Selena Wang
}
