% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simY_nrm.R
\name{simY_nrm}
\alias{simY_nrm}
\title{Simulate a normal behavior matrix}
\usage{
simY_nrm(EY, s1)
}
\arguments{
\item{EY}{matrix giving the expected value of the behavior matrix}

\item{s1}{variance}
}
\value{
a N by P matrix
}
\description{
Simulates a normal behavior matrix
}
\author{
Selena Wang
}
