% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X.R
\docType{data}
\name{X}
\alias{X}
\title{Simulated a list of V by V connectivity data}
\format{
List including a list of weighted adjacency matrices of length N.
}
\usage{
X
}
\description{
The dataset contains a list of V by V connectivity data
}
\details{
a list of V by V connectivity data
}
\keyword{datasets}
