% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Y.R
\docType{data}
\name{Y}
\alias{Y}
\title{Simulated a matrix of N by P individual outcomes data}
\format{
List including a matrix of N by P individual outcomes data about N individuals and P variables measuring a latent construct
}
\usage{
Y
}
\description{
The dataset contains a matrix of N by P individual outcomes data
}
\details{
a matrix of N by P individual outcomes data
}
\keyword{datasets}
